update format and packages (#86)

+++ contracts/lib/Govern.sol
@@ -1,10 +1,10 @@
 pragma solidity ^0.5.0;
 
-import "openzeppelin-solidity/contracts/math/SafeMath.sol";
-import "openzeppelin-solidity/contracts/token/ERC20/IERC20.sol";
-import "openzeppelin-solidity/contracts/token/ERC20/SafeERC20.sol";
-import "openzeppelin-solidity/contracts/ownership/Ownable.sol";
-import "./interface/IGovern.sol";
+import 'openzeppelin-solidity/contracts/math/SafeMath.sol';
+import 'openzeppelin-solidity/contracts/token/ERC20/IERC20.sol';
+import 'openzeppelin-solidity/contracts/token/ERC20/SafeERC20.sol';
+import 'openzeppelin-solidity/contracts/ownership/Ownable.sol';
+import './interface/IGovern.sol';
 
 /**
  * @title Governance module for DPoS contract
@@ -13,23 +13,23 @@ import "./interface/IGovern.sol";
  * @dev Some specific functions of governance are defined in DPoS contract
  */
 contract Govern is IGovern, Ownable {
-    using SafeMath for uint;
+    using SafeMath for uint256;
     using SafeERC20 for IERC20;
 
     struct ParamProposal {
         address proposer;
-        uint deposit;
-        uint voteDeadline;
-        uint record;
-        uint newValue;
+        uint256 deposit;
+        uint256 voteDeadline;
+        uint256 record;
+        uint256 newValue;
         ProposalStatus status;
         mapping(address => VoteType) votes;
     }
 
     struct SidechainProposal {
         address proposer;
-        uint deposit;
-        uint voteDeadline;
+        uint256 deposit;
+        uint256 voteDeadline;
         address sidechainAddr;
         bool registered;
         ProposalStatus status;
@@ -38,13 +38,13 @@ contract Govern is IGovern, Ownable {
 
     IERC20 public governToken;
     // parameters
-    mapping(uint => uint) public UIntStorage;
-    mapping(uint => ParamProposal) public paramProposals;
-    uint public nextParamProposalId;
+    mapping(uint256 => uint256) public UIntStorage;
+    mapping(uint256 => ParamProposal) public paramProposals;
+    uint256 public nextParamProposalId;
     // registered sidechain addresses
-    mapping (address => bool) public registeredSidechains;
-    mapping(uint => SidechainProposal) public sidechainProposals;
-    uint public nextSidechainProposalId;
+    mapping(address => bool) public registeredSidechains;
+    mapping(uint256 => SidechainProposal) public sidechainProposals;
+    uint256 public nextSidechainProposalId;
 
     /**
      * @notice Govern constructor
@@ -60,25 +60,23 @@ contract Govern is IGovern, Ownable {
      */
     constructor(
         address _governTokenAddress,
-        uint _governProposalDeposit,
-        uint _governVoteTimeout,
-        uint _slashTimeout,
-        uint _minValidatorNum,
-        uint _maxValidatorNum,
-        uint _minStakeInPool,
-        uint _advanceNoticePeriod
-    )
-        public
-    {
+        uint256 _governProposalDeposit,
+        uint256 _governVoteTimeout,
+        uint256 _slashTimeout,
+        uint256 _minValidatorNum,
+        uint256 _maxValidatorNum,
+        uint256 _minStakeInPool,
+        uint256 _advanceNoticePeriod
+    ) public {
         governToken = IERC20(_governTokenAddress);
 
-        UIntStorage[uint(ParamNames.ProposalDeposit)] = _governProposalDeposit;
-        UIntStorage[uint(ParamNames.GovernVoteTimeout)] = _governVoteTimeout;
-        UIntStorage[uint(ParamNames.SlashTimeout)] = _slashTimeout;
-        UIntStorage[uint(ParamNames.MinValidatorNum)] = _minValidatorNum;
-        UIntStorage[uint(ParamNames.MaxValidatorNum)] = _maxValidatorNum;
-        UIntStorage[uint(ParamNames.MinStakeInPool)] = _minStakeInPool;
-        UIntStorage[uint(ParamNames.AdvanceNoticePeriod)] = _advanceNoticePeriod;
+        UIntStorage[uint256(ParamNames.ProposalDeposit)] = _governProposalDeposit;
+        UIntStorage[uint256(ParamNames.GovernVoteTimeout)] = _governVoteTimeout;
+        UIntStorage[uint256(ParamNames.SlashTimeout)] = _slashTimeout;
+        UIntStorage[uint256(ParamNames.MinValidatorNum)] = _minValidatorNum;
+        UIntStorage[uint256(ParamNames.MaxValidatorNum)] = _maxValidatorNum;
+        UIntStorage[uint256(ParamNames.MinStakeInPool)] = _minStakeInPool;
+        UIntStorage[uint256(ParamNames.AdvanceNoticePeriod)] = _advanceNoticePeriod;
     }
 
     /********** Get functions **********/
@@ -87,7 +85,7 @@ contract Govern is IGovern, Ownable {
      * @param _record the key of this parameter
      * @return the value of this parameter
      */
-    function getUIntValue(uint _record) public view returns (uint) {
+    function getUIntValue(uint256 _record) public view returns (uint256) {
         return UIntStorage[_record];
     }
 
@@ -97,7 +95,11 @@ contract Govern is IGovern, Ownable {
      * @param _voter the voter address
      * @return the vote type of the given voter on the given parameter proposal
      */
-    function getParamProposalVote(uint _proposalId, address _voter) public view returns (VoteType) {
+    function getParamProposalVote(uint256 _proposalId, address _voter)
+        public
+        view
+        returns (VoteType)
+    {
         return paramProposals[_proposalId].votes[_voter];
     }
 
@@ -116,7 +118,11 @@ contract Govern is IGovern, Ownable {
      * @param _voter the voter address
      * @return the vote type of the given voter on the given sidechain proposal
      */
-    function getSidechainProposalVote(uint _proposalId, address _voter) public view returns (VoteType) {
+    function getSidechainProposalVote(uint256 _proposalId, address _voter)
+        public
+        view
+        returns (VoteType)
+    {
         return sidechainProposals[_proposalId].votes[_voter];
     }
 
@@ -126,22 +132,29 @@ contract Govern is IGovern, Ownable {
      * @param _record the key of this parameter
      * @param _value the new proposed value of this parameter
      */
-    function createParamProposal(uint _record, uint _value) public {
+    function createParamProposal(uint256 _record, uint256 _value) public {
         ParamProposal storage p = paramProposals[nextParamProposalId];
         nextParamProposalId = nextParamProposalId.add(1);
         address msgSender = msg.sender;
-        uint deposit = UIntStorage[uint(ParamNames.ProposalDeposit)];
+        uint256 deposit = UIntStorage[uint256(ParamNames.ProposalDeposit)];
 
         p.proposer = msgSender;
         p.deposit = deposit;
-        p.voteDeadline = block.number.add(UIntStorage[uint(ParamNames.GovernVoteTimeout)]);
+        p.voteDeadline = block.number.add(UIntStorage[uint256(ParamNames.GovernVoteTimeout)]);
         p.record = _record;
         p.newValue = _value;
         p.status = ProposalStatus.Voting;
 
         governToken.safeTransferFrom(msgSender, address(this), deposit);
 
-        emit CreateParamProposal(nextParamProposalId.sub(1), msgSender, deposit, p.voteDeadline, _record, _value);
+        emit CreateParamProposal(
+            nextParamProposalId.sub(1),
+            msgSender,
+            deposit,
+            p.voteDeadline,
+            _record,
+            _value
+        );
     }
 
     /**
@@ -151,11 +164,15 @@ contract Govern is IGovern, Ownable {
      * @param _voter the voter address
      * @param _vote the vote type
      */
-    function internalVoteParam(uint _proposalId, address _voter, VoteType _vote) internal {
+    function internalVoteParam(
+        uint256 _proposalId,
+        address _voter,
+        VoteType _vote
+    ) internal {
         ParamProposal storage p = paramProposals[_proposalId];
-        require(p.status == ProposalStatus.Voting, "Invalid proposal status");
-        require(block.number < p.voteDeadline, "Vote deadline reached");
-        require(p.votes[_voter] == VoteType.Unvoted, "Voter has voted");
+        require(p.status == ProposalStatus.Voting, 'Invalid proposal status');
+        require(block.number < p.voteDeadline, 'Vote deadline reached');
+        require(p.votes[_voter] == VoteType.Unvoted, 'Voter has voted');
 
         p.votes[_voter] = _vote;
 
@@ -168,10 +185,10 @@ contract Govern is IGovern, Ownable {
      * @param _proposalId the proposal id
      * @param _passed proposal passed or not
      */
-    function internalConfirmParamProposal(uint _proposalId, bool _passed) internal {
+    function internalConfirmParamProposal(uint256 _proposalId, bool _passed) internal {
         ParamProposal storage p = paramProposals[_proposalId];
-        require(p.status == ProposalStatus.Voting, "Invalid proposal status");
-        require(block.number >= p.voteDeadline, "Vote deadline not reached");
+        require(p.status == ProposalStatus.Voting, 'Invalid proposal status');
+        require(block.number >= p.voteDeadline, 'Vote deadline not reached');
 
         p.status = ProposalStatus.Closed;
         if (_passed) {
@@ -201,18 +218,25 @@ contract Govern is IGovern, Ownable {
         SidechainProposal storage p = sidechainProposals[nextSidechainProposalId];
         nextSidechainProposalId = nextSidechainProposalId.add(1);
         address msgSender = msg.sender;
-        uint deposit = UIntStorage[uint(ParamNames.ProposalDeposit)];
+        uint256 deposit = UIntStorage[uint256(ParamNames.ProposalDeposit)];
 
         p.proposer = msgSender;
         p.deposit = deposit;
-        p.voteDeadline = block.number.add(UIntStorage[uint(ParamNames.GovernVoteTimeout)]);
+        p.voteDeadline = block.number.add(UIntStorage[uint256(ParamNames.GovernVoteTimeout)]);
         p.sidechainAddr = _sidechainAddr;
         p.registered = _registered;
         p.status = ProposalStatus.Voting;
 
         governToken.safeTransferFrom(msgSender, address(this), deposit);
 
-        emit CreateSidechainProposal(nextSidechainProposalId.sub(1), msgSender, deposit, p.voteDeadline, _sidechainAddr, _registered);
+        emit CreateSidechainProposal(
+            nextSidechainProposalId.sub(1),
+            msgSender,
+            deposit,
+            p.voteDeadline,
+            _sidechainAddr,
+            _registered
+        );
     }
 
     /**
@@ -222,11 +246,15 @@ contract Govern is IGovern, Ownable {
      * @param _voter the voter address
      * @param _vote the vote type
      */
-    function internalVoteSidechain(uint _proposalId, address _voter, VoteType _vote) internal {
+    function internalVoteSidechain(
+        uint256 _proposalId,
+        address _voter,
+        VoteType _vote
+    ) internal {
         SidechainProposal storage p = sidechainProposals[_proposalId];
-        require(p.status == ProposalStatus.Voting, "Invalid proposal status");
-        require(block.number < p.voteDeadline, "Vote deadline reached");
-        require(p.votes[_voter] == VoteType.Unvoted, "Voter has voted");
+        require(p.status == ProposalStatus.Voting, 'Invalid proposal status');
+        require(block.number < p.voteDeadline, 'Vote deadline reached');
+        require(p.votes[_voter] == VoteType.Unvoted, 'Voter has voted');
 
         p.votes[_voter] = _vote;
 
@@ -239,10 +267,10 @@ contract Govern is IGovern, Ownable {
      * @param _proposalId the proposal id
      * @param _passed proposal passed or not
      */
-    function internalConfirmSidechainProposal(uint _proposalId, bool _passed) internal {
+    function internalConfirmSidechainProposal(uint256 _proposalId, bool _passed) internal {
         SidechainProposal storage p = sidechainProposals[_proposalId];
-        require(p.status == ProposalStatus.Voting, "Invalid proposal status");
-        require(block.number >= p.voteDeadline, "Vote deadline not reached");
+        require(p.status == ProposalStatus.Voting, 'Invalid proposal status');
+        require(block.number >= p.voteDeadline, 'Vote deadline not reached');
 
         p.status = ProposalStatus.Closed;
         if (_passed) {
